WordClouds for TF of BoW
Word Clouds for term frequency of Bags of Words using law projects from 2016 to 2021 (11/07/21 exactly). We've separated this information by two factors; first, we have four presidential periods because we've had four diferent presidents in Perú from 2016 to 2021; second, we have two congresal periods because we've had two diferent congress in Perú from 2016 to 2021, this periods are precisely equal to the pre pandemic and pandemic periods. We took this decition because in this way we can see the changes on the words most used in each period.